Separate Smithy-Tags into separate Task
So that it is always run before the Jar Task.
